services/notification-service/requirements.txt
flask
redis
kafka-python
flask-jwt-extended